Change input types to float for bill and tip percent

+++ Day 2/Day2_Poject_Tip Calculator.py
@@ -1,7 +1,7 @@
 print("Welcome to the Tip Calculator!!")
 try:
-    total_bill=int(input("Enter the total bill:$"))
-    tip_percent=int(input("Ënter the total percent of bill you wanna give tip off:%"))
+    total_bill=float(input("Enter the total bill:$"))
+    tip_percent=float(input("Ënter the total percent of bill you wanna give tip off:%"))
     split=int(input("With how many people you wanna split the bill:"))
     final_amount=total_bill + (total_bill/100)*tip_percent
     amount_perhead=format(final_amount/split,".2f")
